Replace a couple naked `ruby` calls with `CommandTRuby` wrapper calls
In the time between when Rainux prepared the pull request that added the
lazy-loading feature (commit f175842) and when it got merged, a couple
of new finders were added to Command-T.

This commit updates them to use the new `CommandTRuby` wrapper function
rather than just calling the built-in `ruby` function. Note that this
isn't just about consistency but also correctness, as bypassing the
wrapper could lead to the `$controller` global variable being nil if the
jump or tag finders are the first Command-T function called in a
session.

Signed-off-by: Wincent Colaiuta <win@wincent.com>